CheeseFun1.0/requirements.txt
requests
colorama
playsound
pyyaml
Pillow
time